requirements.txt
PyYAML>=6.0
pandas
numpy
streamlit
plotly
